Move migration logic out of mod.rs
